+++ packages/astro/src/core/ssr/index.ts
@@ -1,6 +1,19 @@
 import type { BuildResult } from 'esbuild';
 import type vite from '../vite';
-import type { AstroConfig, ComponentInstance, GetStaticPathsResult, Params, Props, Renderer, RouteCache, RouteData, RuntimeMode, SSRElement, SSRError } from '../../@types/astro';
+import type {
+	AstroConfig,
+	ComponentInstance,
+	GetStaticPathsResult,
+	GetStaticPathsResultKeyed,
+	Params,
+	Props,
+	Renderer,
+	RouteCache,
+	RouteData,
+	RuntimeMode,
+	SSRElement,
+	SSRError,
+} from '../../@types/astro';
 import type { LogOptions } from '../logger';
 
 import eol from 'eol';
@@ -14,6 +27,7 @@ import { injectTags } from './html.js';
 import { generatePaginateFunction } from './paginate.js';
 import { getParams, validateGetStaticPathsModule, validateGetStaticPathsResult } from './routing.js';
 import { createResult } from './result.js';
+import { assignStaticPaths, ensureRouteCached, findPathItemByKey } from './route-cache.js';
 
 const svelteStylesRE = /svelte\?svelte&type=style/;
 
@@ -52,9 +66,10 @@ async function resolveRenderer(viteServer: vite.ViteDevServer, renderer: string,
 
 	resolvedRenderer.name = name;
 	if (client) resolvedRenderer.source = path.posix.join(renderer, client);
+	resolvedRenderer.serverEntry = path.posix.join(renderer, server);
 	if (Array.isArray(hydrationPolyfills)) resolvedRenderer.hydrationPolyfills = hydrationPolyfills.map((src: string) => path.posix.join(renderer, src));
 	if (Array.isArray(polyfills)) resolvedRenderer.polyfills = polyfills.map((src: string) => path.posix.join(renderer, src));
-	const { url } = await viteServer.moduleGraph.ensureEntryFromUrl(path.posix.join(renderer, server));
+	const { url } = await viteServer.moduleGraph.ensureEntryFromUrl(resolvedRenderer.serverEntry);
 	const { default: rendererSSR } = await viteServer.ssrLoadModule(url);
 	resolvedRenderer.ssr = rendererSSR;
 
@@ -130,16 +145,18 @@ export async function getParamsAndProps({
 	logging,
 	pathname,
 	mod,
+	validate = true,
 }: {
 	route: RouteData | undefined;
 	routeCache: RouteCache;
 	pathname: string;
 	mod: ComponentInstance;
 	logging: LogOptions;
+	validate?: boolean;
 }): Promise<[Params, Props]> {
 	// Handle dynamic routes
 	let params: Params = {};
-	let pageProps: Props = {};
+	let pageProps: Props;
 	if (route && !route.pathname) {
 		if (route.params.length) {
 			const paramsMatch = route.pattern.exec(pathname);
@@ -147,24 +164,27 @@ export async function getParamsAndProps({
 				params = getParams(route.params)(paramsMatch);
 			}
 		}
-		validateGetStaticPathsModule(mod);
-		if (!routeCache[route.component]) {
-			routeCache[route.component] = await (
-				await mod.getStaticPaths!({
-					paginate: generatePaginateFunction(route),
-					rss: () => {
-						/* noop */
-					},
-				})
-			).flat();
+		if (validate) {
+			validateGetStaticPathsModule(mod);
 		}
-		validateGetStaticPathsResult(routeCache[route.component], logging);
-		const routePathParams: GetStaticPathsResult = routeCache[route.component];
-		const matchedStaticPath = routePathParams.find(({ params: _params }) => JSON.stringify(_params) === JSON.stringify(params));
+		if (!routeCache[route.component]) {
+			await assignStaticPaths(routeCache, route, mod);
+		}
+		if (validate) {
+			// This validation is expensive so we only want to do it in dev.
+			validateGetStaticPathsResult(routeCache[route.component], logging);
+		}
+		const staticPaths: GetStaticPathsResultKeyed = routeCache[route.component];
+		const paramsKey = JSON.stringify(params);
+		const matchedStaticPath = findPathItemByKey(staticPaths, paramsKey, logging);
 		if (!matchedStaticPath) {
 			throw new Error(`[getStaticPaths] route pattern matched, but no matching static path found. (${pathname})`);
 		}
-		pageProps = { ...matchedStaticPath.props } || {};
+		// This is written this way for performance; instead of spreading the props
+		// which is O(n), create a new object that extends props.
+		pageProps = Object.create(matchedStaticPath.props || Object.prototype);
+	} else {
+		pageProps = {};
 	}
 	return [params, pageProps];
 }
@@ -184,16 +204,7 @@ export async function render(renderers: Renderer[], mod: ComponentInstance, ssrO
 			}
 		}
 		validateGetStaticPathsModule(mod);
-		if (!routeCache[route.component]) {
-			routeCache[route.component] = await (
-				await mod.getStaticPaths!({
-					paginate: generatePaginateFunction(route),
-					rss: () => {
-						/* noop */
-					},
-				})
-			).flat();
-		}
+		await ensureRouteCached(routeCache, route, mod);
 		validateGetStaticPathsResult(routeCache[route.component], logging);
 		const routePathParams: GetStaticPathsResult = routeCache[route.component];
 		const matchedStaticPath = routePathParams.find(({ params: _params }) => JSON.stringify(_params) === JSON.stringify(params));